src/requirements.txt
# 核心机器学习包
torch==2.8.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.21.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0

# 工具实现
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia>=1.4.0

# 系统工具
tenacity>=8.2.0
tqdm>=4.65.0
pyyaml>=6.0
fire>=0.5.0

# 强化学习相关
gymnasium>=0.28.0

# 其他工具
python-dotenv>=1.0.0
psutil>=5.9.0